+++ src/converters/piral-react-15/src/react15.d.ts
@@ -1,9 +1,0 @@
-declare module 'react-15' {
-  import { ComponentType, Component, createElement } from 'react';
-  export { ComponentType, Component, createElement };
-}
-
-declare module 'react-dom-15' {
-  import { render } from 'react-dom';
-  export { render };
-}